% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\docType{data}
\name{dat}
\alias{dat}
\title{Example Dataset}
\format{
A data frame with 100 rows and 5 variables:
\describe{
\item{betaX}{Effect size of the exposure}
\item{betaXse}{Standard error of the effect size of the exposure}
\item{betaY}{Effect size of the outcome}
\item{betaYse}{Standard error of the effect size of the outcome}
\item{ny}{Sample size of outcome GWAS}
}
}
\source{
The data was simulated/generated for example purposes.
}
\usage{
dat
}
\description{
This dataset contains example data for testing purposes.
}
\keyword{datasets}
