% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MR_em.R
\name{mr_em}
\alias{mr_em}
\title{Mendelian Randomization analysis}
\usage{
mr_em(
  betaY,
  betaX,
  betaYse,
  betaXse,
  ny,
  gwas_p = 0.05,
  gwas_p2 = 5e-08,
  permutn = 0
)
}
\arguments{
\item{betaY}{beta of outcome, recommended to be in standardized scale.}

\item{betaX}{beta of exposure, recommended to be in standardized scale.}

\item{betaYse}{standard error of outcome, recommended to be in standardized scale.}

\item{betaXse}{standard error of exposure, recommended to be in standardized scale.}

\item{ny}{sample size of outcome gwas.}

\item{gwas_p}{p value of association between variants and the latent phenotype (horizontal pleiotropy test)}

\item{gwas_p2}{p value of exposure gwas (SignifThreshold).}

\item{permutn}{number of permutations.}
}
\value{
A list that contains
\item{CausEst}{Estimate of causal effect.}
\item{CausEstSE}{Standard error of causal effect estimate.}
\item{CausEstP}{P-value from the z test for the causal effect.}
\item{SNPP}{P-value from the for the direct causal effect of variant on the the outcome.}
\item{VPI}{index of valid variants.}
}
\description{
This function conducts Mendelian randomization analysis based on latent phenotype of the outcome that explicitly exclude vertical pleiotropy effects and iteratively refines causal relationship  through an Expectation and Maximisation (EM) algorithm.MRvpi takes GWAS summary statistics as inputs to estimate causal effects of one trait on another. We recommend using summary statistics in the standardized scale
}
\details{
None.
}
\keyword{Mendelian}
\keyword{randomization.}
